+++ tests/test_base_scorer.py
@@ -16,10 +16,10 @@ def base_scorer(scope="function"):
 @pytest.fixture()
 def matrix_multiple_var(scope="function"):
     """Embedding matrix for test"""
-    mat1 = torch.tensor([[3, 4], [3, 4]], dtype=torch.float32)
-    mat2 = torch.tensor([[7, 24], [7, 24], [7, 24]], dtype=torch.float32)
+    mat1 = torch.tensor([[3, 4], [3, 4]], dtype=torch.float64)
+    mat2 = torch.tensor([[7, 24], [7, 24], [7, 24]], dtype=torch.float64)
     expected = torch.tensor(
-        [[0.9360, 0.9360, 0.9360], [0.9360, 0.9360, 0.9360]], dtype=torch.float32
+        [[0.9360, 0.9360, 0.9360], [0.9360, 0.9360, 0.9360]], dtype=torch.float64
     )
 
     return {"mat1": mat1, "mat2": mat2, "expected": expected}
